move asynccontentview
