+++ Snabble/UI/DynamicStack/WidgetPurchaseView.swift
@@ -183,33 +183,3 @@ public struct WidgetPurchaseView: View {
         }
     }
 }
-
-enum LoadingState<Value> {
-    case idle
-    case loading
-    case failed(Error)
-    case loaded(Value)
-    case empty
-}
-
-protocol LoadableObject: ObservableObject {
-    associatedtype Output
-    var state: LoadingState<Output> { get }
-    func load()
-}
-
-struct AsyncContentView<Source: LoadableObject, Content: View>: View {
-    @ObservedObject var source: Source
-    var content: (Source.Output) -> Content
-
-    var body: some View {
-        switch source.state {
-        case .idle:
-            Color.clear.onAppear(perform: source.load)
-        case .loading, .failed, .empty:
-            EmptyView()
-        case .loaded(let output):
-            content(output)
-        }
-    }
-}
